requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_pytorch_select=0.2=gpu_0
asn1crypto=1.3.0=py36_0
blas=1.0=mkl
ca-certificates=2020.1.1=0
cairocffi=1.1.0=pypi_0
cairosvg=2.4.2=pypi_0
certifi=2020.4.5.1=py36_0
cffi=1.12.3=py36h2e261b9_0
chardet=3.0.4=py36_1003
click=7.1.1=py_0
cryptography=2.8=py36h1ba5d50_0
cssselect2=0.3.0=pypi_0
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
cymem=1.31.2=py36h6bb024c_0
cytoolz=0.9.0.1=py36h14c3975_1
defusedxml=0.6.0=pypi_0
dill=0.2.9=py36_0
dominate=2.4.0=py_0
flask=1.1.1=py_1
flask-session=0.3.1=pypi_0
freetype=2.9.1=h8a8886c_1
idna=2.9=py_1
intel-openmp=2019.3=199
itsdangerous=1.1.0=py36_0
jinja2=2.11.1=py_0
joblib=0.14.1=py_0
jpeg=9b=h024ee3a_2
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.1.0=h2733197_0
markupsafe=1.1.1=py36h7b6447c_0
mkl=2019.4=243
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.0.14=py36ha843d7b_0
mkl_random=1.0.2=py36hd81dba3_0
msgpack-numpy=0.4.4.3=py_0
msgpack-python=0.5.6=py36h6bb024c_1
murmurhash=0.28.0=py36hf484d3e_0
nccl=1.3.5=cuda10.0_0
ncurses=6.1=he6710b0_1
ninja=1.9.0=py36hfd86e86_0
nltk=3.4.5=py36_0
numpy=1.16.4=py36h7e9f1db_0
numpy-base=1.16.4=py36hde5b4d6_0
olefile=0.46=py36_0
opencv3=3.1.0=py36_0
openssl=1.1.1g=h7b6447c_0
pandas=1.0.3=py36h0573a6f_0
pillow=6.0.0=py36h34e0f95_0
pip=19.1.1=py36_0
plac=0.9.6=py36_0
preshed=1.0.1=py36he6710b0_0
pycparser=2.19=py36_0
pyopenssl=19.1.0=py36_0
pysocks=1.7.1=py36_0
python=3.6.8=h0371630_0
python-dateutil=2.8.1=py_0
python_abi=3.6=1_cp36m
pytorch=1.0.1=cuda100py36he554f03_0
pytz=2019.3=py_0
readline=7.0=h7b6447c_5
regex=2018.07.11=py36h14c3975_0
requests=2.23.0=py36_0
scikit-learn=0.21.1=py36hd81dba3_0
scipy=1.2.1=py36h7c811a0_0
setuptools=41.0.1=py36_0
six=1.12.0=py36_0
spacy=2.0.12=py36h962f231_0
sqlite=3.28.0=h7b6447c_0
svgwrite=1.4=pypi_0
termcolor=1.1.0=py36_1
thinc=6.10.3=py36h962f231_0
tinycss2=1.0.2=pypi_0
tk=8.6.8=hbc83047_0
toolz=0.10.0=py_0
torchvision=0.2.1=py36_0
tqdm=4.32.1=py_0
ujson=1.35=py36h14c3975_0
urllib3=1.25.8=py36_0
webencodings=0.5.1=pypi_0
werkzeug=0.16.0=pypi_0
wheel=0.33.4=py36_0
wrapt=1.10.11=py36h14c3975_2
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
